[Refactor][Kurt] - Refactored code
