+++ src/app/me/about/name-display.tsx
@@ -1,0 +1,14 @@
+import { Text } from '@mantine/core';
+import React from 'react';
+
+function NameDisplay() {
+  return (
+    <div className='w-[36rem] mt-12 flex flex-col max-md:w-full px-12 text-white'>
+      <Text className='font-extrabold text-[3rem]'>Kurt</Text>
+      <Text className='font-extrabold text-[3rem]'>Jacob</Text>
+      <Text className='font-extrabold text-[3rem]'>Urquico</Text>
+    </div>
+  );
+}
+
+export default NameDisplay;
